Part 2 Done
The boot camp version did it a little different compared to mine. It added a regex to check and remove special characters. It also stored the reversed version of the word and the special message in an object. This would be the first time in this boot camp that an object was used.